librlist: make rlist.o usable without libhwloc
Problem: the sched-simple and job-list broker modules currently
have a dynamic dependency on libhwloc through librlist, but do not
use any functions in librlist that require libhwloc.

To allow these modules to drop the hwloc dependency, make rlist.lo
independent of rhwloc.lo by splitting out rlist_from_hwloc() to
another compilation unit.

Then, if desired, the modules can link directly with rlist.lo,
rnode.lo, and match.lo and avoid hwloc.
