+++ src/common/librlist/rlist_private.h
@@ -1,0 +1,17 @@
+/************************************************************\
+ * Copyright 2014 Lawrence Livermore National Security, LLC
+ * (c.f. AUTHORS, NOTICE.LLNS, COPYING)
+ *
+ * This file is part of the Flux resource manager framework.
+ * For details, see https://github.com/flux-framework.
+ *
+ * SPDX-License-Identifier: LGPL-3.0
+\************************************************************/
+
+#ifndef HAVE_SCHED_RLIST_PRIVATE_H
+#define HAVE_SCHED_RLIST_PRIVATE_H 1
+
+// needed by rlist_hwloc.c
+int rlist_add_rnode (struct rlist *rl, struct rnode *n);
+
+#endif /* !HAVE_SCHED_RLIST_H */
